Initial commit
The motivation behind this project is the book ‘Ansible for DevOps’ by
Jeff Geerling [1].  In one of its chapters, the author describes how one
can use Ansible to automate the configuration of their personal
computers.

I have used Ansible off and on in the past, but never to this extent.
Hence, I intend to construct the configuration step by step, leveraging
this opportunity to take a deep dive into Ansible.

An Ansible playbook is a YAML file.  Its root is a sequence of plays and
each play is a mapping.  The only key that absolutely must exist in a
play is ‘hosts’; it specifies the targets of the play.  The special
value ‘all’ will match all hosts.  The current content of ‘T495s.yml’
constitutes the most minimal YAML file that is a valid playbook.

The following command is used to run the playbook:

    $ ansible-playbook --connection local --inventory T495s, T495s.yml

Typically, Ansible is used to configure remote hosts.  The connection to
them is done using SSH (one of Ansible’s greatest strengths is its
agentless approach).  By using the ‘local’ connection plugin, the
playbook is run on the local environment instead.

An inventory is an organisation of the managed hosts.  However, when
using the ‘local’ connection plugin, all that is needed is a single host
(its name is actually inconsequential). The command-line argument that
follows ‘--inventory’ can be either a path to a file or a
comma-separated list of hosts.

Under the assumption that an Ansible configuration is considered
software, I have chosen the GNU General Public Licence.  The content of
‘LICENCE.md’ was copied from the GNU website [2].

[1] https://leanpub.com/ansible-for-devops
[2] https://www.gnu.org/licenses/gpl-3.0.md