Initial commit
The motivation behind this project is the book ‘Ansible for DevOps’ by
Jeff Geerling [1]. In one of its chapters, the author describes how one
can use Ansible to automate the configuration of their personal
computers.

I have used Ansible off and on in the past, but never to this extent.
Having recently purchased a new laptop, I intend to construct its
configuration step by step, leveraging this opportunity to take a deep
dive into Ansible.

An Ansible playbook is a YAML file. Its root is a sequence of plays and
each play is a mapping. The documentation [2] describes the available
keys. The current content of ‘T14s.yaml’ constitutes the most minimal
YAML file that is a valid playbook.

The following command is used to run the playbook:

    $ ansible-playbook --connection local --inventory localhost, T14s.yaml

Typically, Ansible is used to configure remote hosts. The connection to
them is done using SSH (one of Ansible’s greatest strengths is its
agentless approach). By using the ‘local’ connection plugin, the
playbook is run on the local device instead.

An inventory is an organisation of the managed hosts. However, when
using the ‘local’ connection plugin, all that is needed is a single host
(its name is actually inconsequential). The command-line argument that
follows ‘--inventory’ can be a path to a file or a comma-separated list
of hosts.

Under the assumption that an Ansible configuration is considered
software, I have chosen the GNU General Public Licence. The content of
‘LICENCE.markdown’ was copied verbatim from the GNU website [3].

[1] https://leanpub.com/ansible-for-devops
[2] https://docs.ansible.com/ansible/10/reference_appendices/playbooks_keywords.html#play
[3] https://www.gnu.org/licenses/gpl-3.0.md