playbook: Configure hostname
A task is the most basic unit of execution. Typically, it contains the
description of a desired state in a declarative manner. Ansible will
execute the module used by the task. In turn, the module will execute
all, if any, steps necessary to reach that desired state.

Each task is expressed as a mapping. One of its keys must be the name of
the module to execute. Its value should be a mapping containing the
module’s parameters.

The ‘ansible.builtin.hostname’ module [1] uses ‘hostnamectl’ [2]. Its
effect is both immediate and persistent (the latter by writing the
hostname in ‘/etc/hostname’).

As a side note, it is possible to execute modules in an ad-hoc manner.
The following command is equivalent:

    $ ansible --module-name ansible.builtin.hostname --args name=T14s localhost

[1] https://docs.ansible.com/ansible/10/collections/ansible/builtin/hostname_module.html
[2] https://www.freedesktop.org/software/systemd/man/255/hostnamectl.html