hostname: Add role
Ansible roles offer a structured mechanism to group content (e.g. tasks)
in a reusable way. By default, roles are found as subdirectories inside
a ‘roles’ directory.

A role’s entry point is the file ‘tasks/main.yaml’. Its root is a
sequence of tasks, similar to a playbook’s ‘tasks’ keyword. The
following command can be used to populate a new role with the
recommended structure:

    $ ansible-galaxy role init roles/ROLE_NAME

It might seem excessive to introduce a role with a single task, but
ideally I would like the playbook to only import roles and not contain
any tasks of its own.