hostname: Use a variable
It is recommended to use variables [1] instead of embedding any sort of
data in tasks. They can then be referenced using a Jinja syntax. The
file ‘defaults/main.yaml’ is read automatically. Its root is a mapping
of variables and their values.

In order to avoid unintentional clashes, variables should have the name
of the role followed by an underscore as a prefix [2].

[1] https://docs.ansible.com/ansible/10/playbook_guide/playbooks_variables.html
[2] https://ansible.readthedocs.io/projects/lint/rules/var-naming/