flatpak: Add Flathub repository
Flatpak is the recommended mechanism to install graphical applications
on a Fedora atomic desktop. Flathub is the de facto Flatpak repository,
containing both free and proprietary software from a wide range of
vendors.

The ‘community.general.flatpak_remote’ module can be used to add or
remove repositories. Further capabilities are limited (e.g. it won’t
update an existing repository). The configuration was adapted from the
official documentation [1].

Like ‘defaults/main.yaml’, the file ‘vars/main.yaml’ is read
automatically. The difference is in their precedence [2]. In practical
terms, variables that a user of a role may want to change should be
placed in the former, whereas variables that act more like constants
should go to the latter. It is recommended to prefix such variables with
an underscore.

[1] https://flathub.org/setup/Fedora
[2] https://docs.ansible.com/ansible/10/playbook_guide/playbooks_variables.html#understanding-variable-precedence