requirements: Add file
Commit 885ceae makes use of the ‘community.general’ collection. This may
not be available in a minimal Ansible installation.

The file ‘requirements.yaml’ is used to list all required collections
and roles [1]. Its root should be a mapping with only two possible keys:
‘collections’ and ‘roles’. Their values can be either strings or, for
mode advanced configuration, mappings.

The following command is used to install the requirements:

    $ ansible-galaxy collection install --requirements-file requirements.yaml

Unfortunately, it does not appear possible to make this part of the
playbook.

[1] https://docs.ansible.com/ansible/latest/galaxy/user_guide.html#installing-roles-and-collections-from-the-same-requirements-yml-file