flatpak: Check installed applications from the Fedora repository
Many pre-installed graphical applications use Flatpak. However, they are
built from RPMs, use OCI images, use a Fedora runtime, and follow
Fedora’s release schedule. Instead, applications from Flathub are closer
to upstream releases.

I prefer to uninstall the default selection of applications from Fedora
and install the ones I use from Flathub. The ‘community.general.flatpak’
module can’t do this on its own, so it has to be implemented in two
steps. The first is to identify the installed applications.

The ‘ansible.builtin.shell’ module can be used to implement tasks that
are not possible through existing modules. It offers all the features of
a shell (e.g. redirections, pipelines).

The ‘>-’ YAML syntax is a block scalar. Effectively, the block style
indicator ‘>’ means that internal newlines are converted to spaces and
the block chomping indicator ‘-’ means that the ending newline is
stripped.

Ansible modules return information by writing a JSON string to standard
output. It is possible to store this information and make it available
for subsequent tasks using the ‘register’ keyword. Its value is a string
that must follow the variable naming guidelines [1].

For future reference, the following applications are pre-installed on
Fedora Silverblue 40:

    $ flatpak list --system --app --columns origin,application
    Origin          Application ID
    fedora          org.fedoraproject.MediaWriter
    fedora          org.gnome.Calculator
    fedora          org.gnome.Calendar
    fedora          org.gnome.Characters
    fedora          org.gnome.Connections
    fedora          org.gnome.Contacts
    fedora          org.gnome.Evince
    fedora          org.gnome.Extensions
    fedora          org.gnome.Logs
    fedora          org.gnome.Loupe
    fedora          org.gnome.Maps
    fedora          org.gnome.NautilusPreviewer
    fedora          org.gnome.Snapshot
    fedora          org.gnome.TextEditor
    fedora          org.gnome.Weather
    fedora          org.gnome.baobab
    fedora          org.gnome.clocks
    fedora          org.gnome.font-viewer

[1] https://docs.ansible.com/ansible/10/playbook_guide/playbooks_variables.html#creating-valid-variable-names