flatpak: Treat command as read-only
The ‘ansible.builtin.shell’ module cannot ascertain on its own whether
the command it executes is modifying the target host. Consequently, it
skips the task in check mode [1] and assumes it always makes changes.

The ‘check_mode’ keyword can override the former. Its value is a boolean
that, when false, will force the task to be executed.

The ‘changed_when’ keyword can evaluate the latter. Typically, its value
is a string that must contain a valid raw Jinja expression [2], but a
boolean is also acceptable.

[1] https://docs.ansible.com/ansible/10/playbook_guide/playbooks_checkmode.html
[2] https://jinja.palletsprojects.com/en/3.1.x/templates/#expressions