firefox: Install Firefox
Fedora Silverblue comes with Firefox pre-installed as an RPM package in
its image. However, it is lacking support for patent-encumbered codecs
(e.g. H.264). Typically, this is addressed by enabling the RPM Fusion
repository and installing the necessary packages.

On atomic desktops, a preferable solution is to install it from Flathub.
This offers numerous advantages, such as not having to layer packages on
top of the system image, being able to update independently, and
leveraging the features of Flatpak (e.g. the sandbox).

Both the RPM and Flatpak installations name their Firefox desktop
entries the same. The ‘XDG_DATA_DIRS’ environmental variable, set by
‘/etc/profile.d/flatpak.sh’, dictates the order of precedence. As a
consequence, the Flatpak installation effectively replaces the RPM.