firefox: Install the FFmpeg runtime extension
This provides support for patent-encumbered codecs. Firefox does not
require it as a dependency during installation, but will make use of it
when available.

When there is more than one branch of a Flatpak runtime, it is necessary
to specify which one to install. I have added an additional task to
check the one that Firefox is currently using.

It should be noted that runtime branches that are no longer in use will
not be uninstalled by this role.