firefox: Rewrite task as a command
It would appear that the ‘community.general.flatpak’ module does not
handle identifier triples properly. I have chosen to rewrite it as a
command instead that is skipped in check mode.

The ‘ansible.builtin.command’ module similar to but more restrictive
than ‘ansible.builtin.shell’. It is preferable when no shell features
are necessary.