firefox: Create the configuration directories
The configuration of Firefox is organised in profiles [1]. This poses
some difficulties when it comes to managing it with Ansible. Thankfully,
it is possible to leverage some features intended for enterprise users.

Flatpak supports the concept of unmaintained extensions [2]. In short,
it offers a mechanism to mount a directory from the host system on a
predetermined location in the application sandbox with ease.

The file ‘policies/policies.json’ will apply Firefox policies [3]. Any
configuration not covered by them can be set using Javascript files
located in the directory ‘defaults/pref’ (the functionality appears to
be similar to AutoConfig [4]).

This commit solely creates the necessary directories. This is done using
the ‘ansible.builtin.file’ module. The ‘loop’ keyword will result in
multiple executions of a module within the same task [5]. Its value is a
sequence. In each iteration of the loop, the sequence entry is assigned
to the ‘item’ variable. The ‘ansible.builtin.dirname’ filter acts
similarly to the ‘dirname’ command.

[1] https://support.mozilla.org/en-US/kb/profiles-where-firefox-stores-user-data
[2] https://docs.flatpak.org/en/latest/extension.html#unmaintained-flatpak-extensions
[3] https://support.mozilla.org/en-US/kb/customizing-firefox-using-policiesjson
[4] https://support.mozilla.org/en-US/kb/customizing-firefox-using-autoconfig
[5] https://docs.ansible.com/ansible/10/playbook_guide/playbooks_loops.html