firefox: Configure Firefox
The ‘ansible.builtin.copy’ module can be used to copy files to the
target host. When the value of the ‘src’ parameter is a relative path,
Ansible implicitly searches for it [1]. Commit b2e4cff guarantees the
existence of the destination directories. The ‘ansible.builtin.basename’
filter acts similarly to the ‘basename’ command.

Both files are effectively empty. Changes to the configuration have no
effect on a running Firefox instance.

[1] https://docs.ansible.com/ansible/10/playbook_guide/playbook_pathing.html