firefox: Remove ‘prefs.js’
The Firefox policies have proved to be sufficient for my needs.
Removing the empty file ‘prefs.js’ simplifies the role.

This commit does not undo the effects of the previous configuration. The
remnant file and directories had to be removed manually. The typical way
to undo a task is to replace it with its reverse (e.g. with state
‘absent’).