firefox: Use an equal sign to attach arguments
Though functionally identical, I intend to switch to this format from
now on. When using a space, it is unclear whether the argument is
attached to a command-line option or is a positional argument. In cases
where the argument is optional (e.g. ‘ls --color=auto’), the use of the
equal sign is mandatory.

The history behind command-line options is interesting. The use of two
hyphens, long keywords, and the equal sign, is known as the GNU style.