firefox: Add role dependencies
The configuration of the Flathub repository, which is needed to install
Firefox, is done in the role ‘flatpak’. Hence, that role must be run
before this one. Up until now, the order of the roles in the playbook
enforced this, but there is a better means.

Role dependencies [1] can be added inside a metadata file. Its root is a
mapping. The value of the ‘dependencies’ key is a sequence of either
strings or, for more advanced configuration, mappings.

[1] https://docs.ansible.com/ansible/10/playbook_guide/playbooks_reuse_roles.html#using-role-dependencies