requirements: Add file
Ansible can be installed through PyPI. There are two available
distributions [1]. ‘ansible-core’ is more minimal but sufficient for my
needs. The following command can be used:

    $ pip3 install --requirement=requirements.txt

The benefit of using a requirements file is that more Python packages
can be added as necessary. It should not be confused with the file
‘requirements.yaml’ which is used to install additional collections
through Ansible Galaxy.

[1] https://docs.ansible.com/ansible/10/installation_guide/intro_installation.html#selecting-an-ansible-package-and-version-to-install