gnome: Configure GNOME
The GNOME desktop and its applications use dconf to store their
preferences. By default, the dconf database is in binary format. The
‘community.general.dconf’ module can be used to manage it. It depends on
the ‘psutil’ and ‘PyGObject’ Python packages.

The empty mapping that is introduced will be gradually populated in
future commits. The ‘ansible.builtin.dict2items’ filter converts it into
a sequence, which is necessary to implement a loop.